fix: Move OneDimensionalGrid's 'new' function to its own impl block
OneDimensionalGrid's 'new' function relied on the D type parameter which is bound by the Default trait. This collided with the unbound T type parameter declared by the impl block, which forced users of the function to specify an arbitrary 'T' type, unrelated to the actual type 'D' which is used.

To solve this, the function was moved to a new impl block, whose type parameter to begin with was 'D'