refactor: Simplify SimulationGrid's generic parameters
Replace SimulationGrid's generic type parameter T with an associated type (the data the grid holds). Additionally, tie the reference returned by get_value_at to the lifetime of the grid (instead of a generic lifetime).

These two changes clarify the trait and ease its use. The effects are already seen in the SimulationData enum