feat: Create adapters module and an AdapterFactory
The adapters module was created to allow multiple different adapter implementations. In addition to each being a module in the adapters module, the module offers an AdapterFactory enum, which abstracts the creation of the adapters while allowing the backend to use a concrete type as an adapter, instead of a dynamic adapter object.

The factory currently cannot create anything since no adapter is implemented, but that will be handled later