refactor: Replace TryFrom trait bound on SimulationDataAdapter with a function
Remove the TryFrom trait bound which was placed on the SimulationDataAdapter trait with a function that serves the same purpose - from_simulation_data.

This refactoring stems mainly from the TryFrom trait's Error type argument, which conflicts with the AdapterError type argument declared in SimulationDataAdapter. To ensure AdapterError is used, TryFrom is removed