refactor: Move grid_state field from EngineOutput into SimulationEngine
The cells' state represents the cells material - an attribute that throughout the simulation will not change as frequently as the velocity field. Therefor, sending it again and again in the engine's output is wasteful. The state should be in the engine and not shared every iteration.

The only time it can change is upon frontend events (for example if we choose to add obstacles the user can move). In such case, the frontend should send some kind of message to the backend, which will update the state grid accordingly.

Only downside - some adapters may want to know it. For this reason, the grid state will probably move again to the main backend struct in the future