refactor: Move grid dimension checking out of apply_gravity and into compute_timestep_internal
Other functions in the computation will likely want to ensure the grid dimensions match as well. Therefor, it is easier to check it once, then call the rest of the functions, rather than checking it in every possible function.

The apply_gravity function now uses 'expect' instead of propagating errors, since it will not ensure the dimensions from now on