feat: Implement projection in the engine
Implement the projection step in the simulation, which ensures the fluid is incompressible.

The step adds two configurable fields - projection_iterations (the number of times velocity corrections will happen), and overrelaxation_factor (a scalar value which will be multiplied by the divergence to speed up the projection solver). Currently they are set to hard-coded constant values, but in the future they should be fully configurable