HUGE refactor: Change entire engine in preparation for PIC/FLIP method
Change the SimulationEngine struct and its functions dramatically. We will use particles instead of only staggered grids in the simulation, which makes the previous implementation of gravity useless (which is why it is deleted).

In addition to that, the engine will store two staggered velocity grids, but they will not be used in the pool. Instead, the pool will store particles